issue-73 Review Comments incorporated
